Created libtest, improved test harness
* libtest is a library which allows you to create and run tests through
a UART interface. Each test consists of a name and a function to
execute. The tests run on the 32M1. When the micro receives a COUNT
command, it writes the number of tests it plans to execute to UART. When
the micro received a START command, it begins the current test. When a
test is complete, the micro sends a DONE message. In the future, the
goal is to add additional structure to this framework, for example,
ASSERT commands, which ask the client to compare two values.

* The test harness, harness.py, is a python script which automatically
builds tests for the 32M1, uploads them to the specified board, and
captures test output. See the examples/test_suite and tests directory
for examples and more information.