Support multiboard testing, timed assertions
* Multiboard tests use a master/slave configuration. The master test
suite determines the number of tests to run, and is responsible for
initiating test execution. Test suites must be named main1.c, main2.c
etc. The file main1.c corresponds to the master test suite. The test
harness now takes as input a list of ports via the -p option. The master
test suite is uploaded to the board attached to the first port. More
generally, the i-th main.c file is uploaded to the board attached to the
i-th port in the list of input ports. Both the slave and the master
suites are capable of writing to the UART client, and thus, can perform
assertions and print output.

* Each test_t struct now has an optional time member, which indicates
approximately how much time the test should take. If the time taken to
complete the test differs significantly from this value, an error is
thrown.

* Created a simple multiboard test. See the tests folder for details.