Added optional argument to specify UART port(s)
* The optional --uart/-u argument accepts a list of UART ports. The
first UART port is associated with the first programming port (supplied
with the --prog/-p argument), and more generally, the i-th UART port is
associated with the i-th programming port. The number of UART ports
supplied must match the number of programming ports supplied. If the
--uart/-u argument is not specified, the UART port corresponding to each
programming port is guessed. On Windows, this is done by subtracting one
from the corresponding programming port, e.g. if COM6 is provided as the
programming port, COM5 is used as the UART port. On macOS, this is done
by adding 2 to the corresponding programming port, e.g. if
/dev/tty.usbmodem00208212 is provided as the programming port,
/dev/tty.usbmodem00208214 is used as the UART port.

* The test harness script can now be run directly from the command line.

* Improved help and usage messages.

* Added Python 3 support.