Created ADC library
* The analog-to-digital converter (ADC) converts voltages across pins
(also called channels) into digital readings. The voltage measurements
are made relative to ground. The ADC operates in two modes: manual and
auto. Manual mode is used to poll individual channels, whereas auto mode
is used to poll a sequence of channels. Operating in auto mode is more
efficient than operating in manual mode when there are many channels to
poll.

* The adc_t struct associated to each ADC encapsulates information about
its pinout and which channels to poll when in auto mode. The pin_info_t*
cs field dictates the chip select pinout. The uint16_t channel field is
a bitstring; channel j will be polled in auto mode if and only if the
j-th bit of the channel field is 1.

* Created a pin_info_t struct in spi.h to encapsulate information about
device pin configurations.

* Created a conversions library to house functions which convert between
raw ADC readings and actual measurements.

* Created func_tests folder to house functional tests that cannot be
executed through the test harness.