Rewrote queue, stack internals
* The size fields of both structs have been replaced by on-the-fly index
computations.

* The dequeue() and pop() functions now accept NULL as a second argument
instead of an array of uint8_t's. This has the effect of throwing away
the data contained in the dequeued/popped element. This is useful if you
don't care about this data, or already know what it is.

* The dequeue() and pop() functions now erase the memory contents of the
dequeued/popped element.

* Successful queue and stack operations now return 1 instead of 0. This
makes for more idiomatic code.

* Both the queue and stack is_full() and is_empty() functions have been
renamed to queue_full(), stack_full() and queue_empty(), stack_empty(),
respectively.

* Improved the enqueue() and push() function signatures to communicate
that the data passed in will not be modified.

* Minor style fixes.